+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.1.3] - 2025-03-26
+### Changed
+- Remove any failed CSS links from the DOM
+
+
+## [0.1.2] - 2025-03-26
+### Changed
+- Changed developer path option into an autoload that finds the script path.
+
+### Fixed
+- Fixed the path to the npm script in the package.json file.
+
+
 ## [0.1.1] - 2025-03-26
 ### Changed
 - Let the standard paths be a fallback of npm paths. So if the path to npm fails, it will try the standard path.